+++ R/setColours.R
@@ -9,9 +9,12 @@
 #' by the colour you specify. Colours for names that did not yet have a colour
 #' will be appended to the list of colours.
 #'
-#' Do not use this for setting the colours or linetypes of species, because
-#' those are determined by setting the `linecolour` and `linetype` variables in
-#' the species parameter data frame.
+#' If a name coincides with the name of a species, the `linecolour` (for
+#' `setColours()`) or `linetype` (for `setLinetypes()`) entry for that species
+#' in `species_params` and `given_species_params` is updated as well, so that
+#' the choice persists with the species. Alternatively you can set the
+#' `linecolour` and `linetype` variables in the species parameter data frame
+#' directly, see the example below.
 #'
 #' You can use the same colours in your own ggplot2 plots by adding
 #' `scale_colour_manual(values = getColours(params))` to your plot. Similarly
@@ -26,8 +29,9 @@
 #' @examples
 #' params <- setColours(NS_params, list("Resource" = "red","Total" = "#0000ff"))
 #' params <- setLinetypes(NS_params, list("Total" = "dotted"))
-#' # Set colours and linetypes for species
-#' species_params(params)["Cod", "linecolour"] <- "black"
+#' # Set colours and linetypes for species, either via setColours()/
+#' # setLinetypes() or directly via the species parameter data frame
+#' params <- setColours(params, list("Cod" = "black"))
 #' species_params(params)["Cod", "linetype"] <- "dashed"
 #' plotSpectra(params, total = TRUE)
 #' getColours(params)
@@ -44,6 +48,8 @@ setColours.MizerParams <- function(params, colours) {
     params@linecolour <- unlist(
         modifyList(as.list(params@linecolour), colours))
 
+    params <- sync_species_plot_column(params, colours, "linecolour")
+
     params@time_modified <- lubridate::now()
     params
 }
@@ -57,6 +63,34 @@ getColours <- function(params) {
 #' @export
 getColours.MizerParams <- function(params) {
     params@linecolour
+}
+
+# If any of the given names are species names, update the corresponding
+# entry in `species_params` and `given_species_params` so that the choice
+# persists with the species rather than living only in the plotting slot.
+sync_species_plot_column <- function(params, values, column) {
+    species_values <- values[names(values) %in% params@species_params$species]
+    if (length(species_values) == 0) {
+        return(params)
+    }
+    idx <- match(names(species_values), params@species_params$species)
+    new_vals <- unlist(species_values)
+
+    old_vals <- params@species_params[[column]]
+    if (is.null(old_vals)) {
+        old_vals <- rep(NA_character_, nrow(params@species_params))
+    }
+    old_vals[idx] <- new_vals
+    params@species_params[[column]] <- old_vals
+
+    old_given <- params@given_species_params[[column]]
+    if (is.null(old_given)) {
+        old_given <- rep(NA_character_, nrow(params@given_species_params))
+    }
+    old_given[idx] <- new_vals
+    params@given_species_params[[column]] <- old_given
+
+    params
 }
 
 validColours <- function(colours) {
@@ -89,6 +123,8 @@ setLinetypes.MizerParams <- function(params, linetypes) {
     }
     params@linetype <- unlist(
         modifyList(as.list(params@linetype), as.list(linetypes)))
+
+    params <- sync_species_plot_column(params, linetypes, "linetype")
 
     params@time_modified <- lubridate::now()
     params